requirements.txt
python-telegram-bot>=20.7
requests>=2.31.0
newspaper3k>=0.2.8
youtube-transcript-api>=0.6.2
pytube>=15.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Security and utilities
pip-audit>=2.7.0